+++ backend/services/forms/internal/adapters/rest/dto/page.go
@@ -7,7 +7,7 @@ import (
 type PageRequest struct {
 	Key      string           `json:"key" validate:"required,max=25"`
 	Name     string           `json:"name" validate:"required,max=75"`
-	Position int              `json:"position" validate:"gte=0,lte=10"`
+	Position float32          `json:"position" validate:"gte=0,lte=10"`
 	Sections []SectionRequest `json:"sections" validate:"dive"`
 	Rules    []RuleRequest    `json:"rules" validate:"dive"`
 }
@@ -16,7 +16,7 @@ type PageResponse struct {
 	ID       domain.PageID      `json:"id"`
 	Key      string             `json:"key"`
 	Name     string             `json:"name"`
-	Position int                `json:"position"`
+	Position float32            `json:"position"`
 	Sections []*SectionResponse `json:"sections"`
 	Rules    []*RuleResponse    `json:"rules"`
 }
